+++ formats/color-map.module.js.js
@@ -1,14 +1,14 @@
-const { generateColorMap } = require('../utils')
+const { generateColorMap } = require('../utils');
 
 module.exports = (result) => {
-  const swatches = generateColorMap(result.toJS())
+  const swatches = generateColorMap(result.toJS());
 
   const colors = Object.entries(swatches)
     .map(([hue, shade]) => `\
 /** Color ${hue} */
 export const ${hue} = ${JSON.stringify(shade, null, 2)}
 `)
-    .join('\n')
+    .join('\n');
 
-  return colors
-}
+  return colors;
+};