+++ formats/color-swatches.dart.js
@@ -1,25 +1,25 @@
-const { getShade } = require('../utils')
-const groupBy = require('lodash/groupBy')
+const { getShade } = require('../utils');
+const groupBy = require('lodash/groupBy');
 
 class ColorSwatches {
   constructor ({ props }) {
     const propsWithPalette = props.reduce((palettes, prop) => {
-      if (prop.type !== 'color') return palettes
+      if (prop.type !== 'color') return palettes;
 
-      const [, color] = prop.name.split('-')
-      prop.palette = color
-      palettes.push(prop)
+      const [, color] = prop.name.split('-');
+      prop.palette = color;
+      palettes.push(prop);
 
-      return palettes
-    }, [])
+      return palettes;
+    }, []);
 
-    this.palettes = groupBy(propsWithPalette, 'palette')
-    this.colors = props.filter(prop => prop.type === 'color')
+    this.palettes = groupBy(propsWithPalette, 'palette');
+    this.colors = props.filter(prop => prop.type === 'color');
   }
 
   renderPalette (palette) {
-    const props = this.palettes[palette]
-    const _palette = palette.replace('spartan_', '')
+    const props = this.palettes[palette];
+    const _palette = palette.replace('spartan_', '');
     return `\
   /// ${_palette} and swatches
   static const SpartanColor ${_palette} = SpartanColor(
@@ -28,7 +28,7 @@ class ColorSwatches {
       ${props.map(prop => `'${getShade(prop.name)}': Color(${prop.value}),`).join('\n      ')}
     },
   );
-  static const int _${_palette}PrimaryValue = ${props.find(prop => getShade(prop.name) === 'base').value};`
+  static const int _${_palette}PrimaryValue = ${props.find(prop => getShade(prop.name) === 'base').value};`;
   }
 
   render () {
@@ -46,11 +46,11 @@ ${Object.keys(this.palettes)
         .map(palettes => this.renderPalette(palettes))
         .join('\n\n')}
 }
-`
+`;
   }
 }
 
 module.exports = function (result) {
-  const swatches = new ColorSwatches(result.toJS())
-  return swatches.render()
-}
+  const swatches = new ColorSwatches(result.toJS());
+  return swatches.render();
+};